Split `index.js` into separate files (#500)
